+++ src/InfluxDB2/WriteOptions.php
@@ -6,14 +6,16 @@ class WriteOptions
 {
     const DEFAULT_BATCH_SIZE = 10;
     const DEFAULT_RETRY_INTERVAL = 1000;
-    const DEFAULT_MAX_RETRIES = 3;
-    const DEFAULT_MAX_RETRY_DELAY = 15000;
+    const DEFAULT_MAX_RETRIES = 5;
+    const DEFAULT_MAX_RETRY_DELAY = 180000;
+    const DEFAULT_EXPONENTIAL_BASE = 5;
 
     public $writeType;
     public $batchSize;
     public $retryInterval;
     public $maxRetries;
     public $maxRetryDelay;
+    public $exponentialBase;
 
     /**
      * WriteOptions constructor.
@@ -24,6 +26,8 @@ class WriteOptions
      *          'maxRetries' => max number of retries when write fails
      *              The retry interval is used when the InfluxDB server does not specify "Retry-After" header.
      *          'maxRetryDelay' => maximum delay when retrying write
+     *          'exponentialBase' => the base for the exponential retry delay, the next delay is computed as
+     *              `retry_interval * exponential_base^(attempts - 1) + random(jitter_interval)`
      *      ]
      * @param array $writeOptions Array containing the write parameters (See above)
      */
@@ -35,6 +39,7 @@ class WriteOptions
         $this->retryInterval = $writeOptions["retryInterval"] ?? self::DEFAULT_RETRY_INTERVAL;
         $this->maxRetries = $writeOptions["maxRetries"] ?? self::DEFAULT_MAX_RETRIES;
         $this->maxRetryDelay = $writeOptions["maxRetryDelay"] ?? self::DEFAULT_MAX_RETRY_DELAY;
+        $this->exponentialBase = $writeOptions["exponentialBase"] ?? self::DEFAULT_EXPONENTIAL_BASE;
     }
 }