+++ tests/WriteApiTest.php
@@ -145,6 +145,10 @@ class WriteApiTest extends BasicTest
             // not called
             new Response(429));
 
+        $this->writeApi->writeOptions->maxRetries = 3;
+        $this->writeApi->writeOptions->maxRetryDelay = 15000;
+        $this->writeApi->writeOptions->exponentialBase = 2;
+
         $point = Point::measurement('h2o')
             ->addTag('location', 'europe')
             ->addField('level', 2);